test/web-platform-tests/to-upstream/dom/nodes/Range-getBoundingClientRect.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Range.prototype.getBoundingClientRect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const zeroDimensions = {
    bottom: 0,
    height: 0,
    left: 0,
    right: 0,
    top: 0,
    width: 0,
    x: 0,
    y: 0
  };

  const r = document.createRange();
  assert_object_equals(r.getBoundingClientRect(), zeroDimensions);

}, "getBoundingClientRect should return 0 value dimensions for a range");
</script>
